refactoring add animals to be add item and be able to be reused
