+++ src/react/animalList/AnimalsPage.tsx
@@ -1,7 +1,9 @@
 import { useEffect, useState } from 'react';
 import {ShowAnimals } from './showAnimals';
-import { AddAnimals } from './addAnimals';
+import { AddItems } from '../addItem/addItems';
 import { Animals } from '../../data/animal.schema';
+
+const animalColumns = ["species", "subspecies", "binomial"];
 
 const AnimalsPage = () => {
   const [showAdd, setshowAdd] = useState(false);
@@ -15,14 +17,18 @@ const AnimalsPage = () => {
     const setList = list ?? window?.electron?.sendSync('get-animals')
     setAnimalList(setList);
   }
+  const addAnimal = (newAnimal: object) => {
+    updateAnimalList(window?.electron?.sendSync('add-animals', newAnimal));
+    togglesetshowAdd();
+  }
 
   useEffect(() => {
     updateAnimalList(); 
   }, [location.href]);
   
   return <>
-      { showAdd ? <AddAnimals onSubmit={updateAnimalList} onToggle={togglesetshowAdd}/> : <button onClick={togglesetshowAdd}>Add Item</button>}
-      <ShowAnimals animalList={AnimalList}/>
+      { showAdd ? <AddItems onSubmit={addAnimal}  columnList={animalColumns}/> : <button onClick={togglesetshowAdd}>Add Item</button>}
+      { AnimalList && <ShowAnimals animalList={AnimalList}/>}
     </>;
 }
